refactor: move progress-predictor to drive/, remove dead re-export
PR #343 review fixes:
- Move src/progress-predictor.ts → src/drive/progress-predictor.ts (Layer 2)
- Remove dead ProgressPredictor re-export from core-loop.ts
- Remove redundant predicted stall guard in applyStallAction (kept in detectStallsAndRebalance)
- Update all import paths
