+++ src/loop/core-loop-phases-b.ts
@@ -188,15 +188,6 @@ async function applyStallAction(
   result: LoopIterationResult,
   logPrefix: string
 ): Promise<void> {
-  // Predicted stall types are early warnings — log but do not pivot/escalate
-  if (
-    stallReport.stall_type === "predicted_plateau" ||
-    stallReport.stall_type === "predicted_regression"
-  ) {
-    ctx.logger?.info(`CoreLoop: early warning ${stallReport.stall_type} — monitoring`, { goalId });
-    return;
-  }
-
   if (ctx.deps.learningPipeline) {
     try {
       await ctx.deps.learningPipeline.onStallDetected(goalId, stallReport);
